#83 [FEAT] 캘린더 api 연결

+++ src/apis/CalendarAPi.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
-
+import { useAuthStore } from "../hooks/useAuthStore";
 const base_URL = import.meta.env.VITE_API_URL;
-const accessToken = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIzIiwidXNlclR5cGUiOiJwYXRpZW50IiwiaWF0IjoxNzYzNzQ4NzUwLCJleHAiOjE3NjM3NTk1NTB9.i40ZELIj7nW9BZZo18JUTIyrk7fSKksQ_zBu_8ZNIRw";
+//const accessToken = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIzIiwidXNlclR5cGUiOiJwYXRpZW50IiwiaWF0IjoxNzYzNzQ4NzUwLCJleHAiOjE3NjM3NTk1NTB9.i40ZELIj7nW9BZZo18JUTIyrk7fSKksQ_zBu_8ZNIRw";
 export interface scheduleDetail {
     period: "morning" | "lunch" | "dinner" | "bedtime" | string;
     time: string;
@@ -25,6 +25,11 @@ export interface dailyRecord {
 
 // 특정 날짜 복약 일정 조회
 export const fetchDailyRecord = async (date: string) => {
+    const accessToken = useAuthStore.getState().accessToken;
+    if (!accessToken) {
+        console.error("accessToken이 없습니다.");
+        throw new Error("Authorization token not found");
+    }
     try {
         const res = await axios.get(
             `${base_URL}/api/patients/medications?date=${date}`, {
@@ -33,7 +38,30 @@ export const fetchDailyRecord = async (date: string) => {
                 }
             }
         )
-        console.log("전송 성공:", res.data);
+        //console.log("전송 성공:", res.data);
+        return await res.data;
+    } catch (error) {
+        console.error("날짜별 복약 일정 조회 실패: ", error);
+        throw error;
+    }
+}
+
+// 특정 날짜 복약 여부 시간 별 조회
+export const fetchDailyTake = async (date: string) => {
+    const accessToken = useAuthStore.getState().accessToken;
+    if (!accessToken) {
+        console.error("accessToken이 없습니다.");
+        throw new Error("Authorization token not found");
+    }
+    try {
+        const res = await axios.get(
+            `${base_URL}/api/patients/medications/daily?date=${date}`, {
+                headers: {
+                    'Authorization': `Bearer ${accessToken}`,
+                }
+            }
+        )
+        //console.log("전송 성공:", res.data);
         return await res.data;
     } catch (error) {
         console.error("날짜별 복약 일정 조회 실패: ", error);
@@ -42,6 +70,12 @@ export const fetchDailyRecord = async (date: string) => {
 }
 
 export const fetchAllTreatment = async (year: string, month: string) => {
+    const accessToken = useAuthStore.getState().accessToken;
+    if (!accessToken) {
+        console.error("accessToken이 없습니다.");
+        throw new Error("Authorization token not found");
+    }
+
     try {
         const res = await axios.get(`${base_URL}/api/patients/records/dates?year=${year}&month=${month}`, {
             headers: {
@@ -58,6 +92,12 @@ export const fetchAllTreatment = async (year: string, month: string) => {
 
 // 특정 날짜 진료 이력 조회
 export const fetchDailyTreatment = async (date: string) => {
+    const accessToken = useAuthStore.getState().accessToken;
+    if (!accessToken) {
+        console.error("accessToken이 없습니다.");
+        throw new Error("Authorization token not found");
+    }
+
     try {
         const res = await axios.get(`${base_URL}/api/patients/records?date=${date}`, {
             headers: {
@@ -79,6 +119,12 @@ export const fetchDailyTreatment = async (date: string) => {
 
 // 복약 일정 추가
 export const postMedication = async (data: medProps) => {
+    const accessToken = useAuthStore.getState().accessToken;
+    if (!accessToken) {
+        console.error("accessToken이 없습니다.");
+        throw new Error("Authorization token not found");
+    }
+
     try {
         const res = await axios.post(`${base_URL}/api/patients/medications`, data, {
             headers: {
@@ -95,8 +141,38 @@ export const postMedication = async (data: medProps) => {
     }
 }
 
+// 복용 일정 수정
+export const patchMedication = async (recordId: number, data: medProps) => {
+    const accessToken = useAuthStore.getState().accessToken;
+    if (!accessToken) {
+        console.error("accessToken이 없습니다.");
+        throw new Error("Authorization token not found");
+    }
+
+    try {
+        const res = await axios.patch(`${base_URL}/api/patients/medications/${recordId}`, data, {
+            headers: {
+                "Authorization": `Bearer ${accessToken}`,
+                "Content-Type": "application/json",
+            }
+        })
+
+        console.log("복약 일정 수정 성공: ", res.data);
+        return res.data;
+    } catch(error) {
+        console.log("복약 일정 수정 실패 : ", error);
+        throw error;
+    }
+}
+
 // 복용 여부 업데이트
 export const updateMed = async (recordId: string, isTaken: boolean, date: string, period: string) => {
+    const accessToken = useAuthStore.getState().accessToken;
+    if (!accessToken) {
+        console.error("accessToken이 없습니다.");
+        throw new Error("Authorization token not found");
+    }
+
     try {
         const res = await axios.patch(`${base_URL}/api/patients/medications/${recordId}/history`, {
             date: date,
@@ -118,6 +194,12 @@ export const updateMed = async (recordId: string, isTaken: boolean, date: string
 
 // 복약 일정 전체 삭제
 export const deleteMedAll = async (recordId: number, date: string) => {
+    const accessToken = useAuthStore.getState().accessToken;
+    if (!accessToken) {
+        console.error("accessToken이 없습니다.");
+        throw new Error("Authorization token not found");
+    }
+
     try {
         const res = await axios.delete(`${base_URL}/api/patients/medications/${recordId}/after?date=${date}`, {
             headers: {
@@ -134,3 +216,23 @@ export const deleteMedAll = async (recordId: number, date: string) => {
 
 
 // 복약 일정 당일만 삭제
+export const deleteMedSingle = async (recordId: number, date: string) => {
+    const accessToken = useAuthStore.getState().accessToken;
+    if (!accessToken) {
+        console.error("accessToken이 없습니다.");
+        throw new Error("Authorization token not found");
+    }
+
+    try {
+        const res = await axios.delete(`${base_URL}/api/patients/medications/${recordId}/single?date=${date}`, {
+            headers: {
+                'Authorization' : `Bearer ${accessToken}`,
+            }
+        });
+        console.log(res.data);
+        return res.data
+    } catch (error) {
+        console.log(error);
+        throw error;
+    }
+}
